Refactor: Extract workspace `openssh_sftp_protocol_error`
